Update migration file to prevent errors in phpunit

+++ src/migrations/2018_07_26_142915_update_events_table_organization_id.php
@@ -27,14 +27,16 @@ class UpdateEventsTableOrganizationId extends Migration
      */
     public function down()
     {
-        Schema::table('events', function(Blueprint $table) {
-            if (Schema::hasColumn('events', 'organization_id')) {
-                $table->dropColumn('organization_id');
-            }
+        if (Config::get('database.default') === 'mysql') {
+            Schema::table('events', function(Blueprint $table) {
+                if (Schema::hasColumn('events', 'organization_id')) {
+                    $table->dropColumn('organization_id');
+                }
 
-            if (Schema::hasColumn('events', 'organization_type')) {
-                $table->dropColumn('organization_type');
-            }
-        });
+                if (Schema::hasColumn('events', 'organization_type')) {
+                    $table->dropColumn('organization_type');
+                }
+            });
+        }
     }
 }
